requirements.txt
molecule
ansible
jinja2
pyyaml
mock